chainwatch_backend/requirements.txt
# requirements.txt - Python Dependencies for ChainWatch Backend

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# HTTP Client
aiohttp==3.9.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML
google-generativeai==0.3.2

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0

# Task Scheduling (if needed for more advanced scheduling)
APScheduler==3.10.4